+++ example/lib/video_page.dart
@@ -21,27 +21,33 @@ class _VideoScreenState extends State<VideoScreen> {
   @override
   void initState() {
     super.initState();
-    player.setDataSource(widget.url, autoPlay: true);
+    player.setDataSource(widget.url, autoPlay: true).catchError((e) {
+      FijkException fe = e as FijkException;
+      //setState(() {
+      //  errorMsg = fe.message;
+      //});
+      print("setDataSource exception: $fe");
+    }, test: (e) => e is FijkException);
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: FijkAppBar.defaultSetting(title: "Video"),
-        body: Container(
+      appBar: FijkAppBar.defaultSetting(title: "Video"),
+      body: Container(
           child: FijkView(
-            player: player,
-            // panelBuilder: simplestUI,
-            // panelBuilder: (FijkPlayer player, BuildContext context,
-            //     Size viewSize, Rect texturePos) {
-            //   return CustomFijkPanel(
-            //       player: player,
-            //       buildContext: context,
-            //       viewSize: viewSize,
-            //       texturePos: texturePos);
-            // },
-          ),
-        ));
+        player: player,
+        // panelBuilder: simplestUI,
+        // panelBuilder: (FijkPlayer player, BuildContext context,
+        //     Size viewSize, Rect texturePos) {
+        //   return CustomFijkPanel(
+        //       player: player,
+        //       buildContext: context,
+        //       viewSize: viewSize,
+        //       texturePos: texturePos);
+        // },
+      )),
+    );
   }
 
   @override