+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ##  next
 --------------------------------
 
+## 0.1.0
+- update ijkplauer to f0.3.5
+- fijkplayer err state and FijkException
+- support playing flutter asset file 
+- unit test and widget test
+- pass fijkoption arguments and set player's option
+
 ## 0.0.9
 --------------------------------
 - update ijkplayer to f0.3.4